creado el abstract connection repository
con este se hacen todas las conecciones a la base de datos
